==> 6_time_trend.ipynb <==
import pandas as pd
import datetime
from scipy import stats
import matplotlib.pyplot as plt
data_type = {'created_at':'string','id':'string','full_text':'string','user':'object'}
data = pd.read_csv("/data1/data_8t/lwx/Drugtweets/药物分布/vitamin d|Vitamin D2|Vitamin D3|vd3|vd2.csv",dtype=data_type, lineterminator='\n',usecols=['id','created_at','user','full_text'],low_memory=False, keep_default_na=False)
case_data=pd.read_csv("/data1/data_8t/lwx/Drugtweets/药物分布/case_day.csv")
#### Generate frequency (freq) files for each of the five drugs separately."
data=pd.read_csv("/data1/data_8t/lwx/Drugtweets/药物分布/vitamin d|Vitamin D2|Vitamin D3|vd3|vd2.csv")
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
del data['created_at']
data.insert(data.shape[1], 'd', 1)
freq=data.resample('W', label='right', closed='right').d.sum()
print(freq)
freq.to_csv("freqvitd.csv",index=True)
#freq.to_csv("case_week.csv",index=True)
import numpy as np
rng = pd.date_range('20180103', periods=12)
ts = pd.Series(np.arange(1,13), index=rng)
ts
# 2018-01-03     1
# 2018-01-04     2
# 2018-01-05     3
# 2018-01-06     4
# 2018-01-07     5
# 2018-01-08     6
# 2018-01-09     7
# 2018-01-10     8
# 2018-01-11     9
# 2018-01-12    10
# 2018-01-13    11
# 2018-01-14    12
#### Plot figures
import numpy as np
import pandas as pd
import matplotlib.dates as mdates    # Handle dates
import matplotlib.pyplot as plt

# Read in dates
fig = plt.figure(figsize=(35, 12))
ax1 = fig.add_subplot()
font1 = {'family' : 'DejaVu Sans',
'weight' : 'normal',
'size'   : 20,
}

df1 = pd.read_csv('freqiver.csv')
x = df1['created_at']
x = pd.to_datetime(x)               # Convert to date, otherwise the date settings below won't take effect
y1 = df1['d']

df2 = pd.read_csv('freqhcq.csv')
x = df2['created_at']
x = pd.to_datetime(x)               # Convert to date, otherwise the date settings below won't take effect
y2 = df2['d']

# plt.gca() function gets the current axis, then parameters or plots can be set. plt.plot() internally implements this step.
plt.gca().xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))  # Set the display format of the x-axis major ticks (date)
plt.gca().xaxis.set_major_locator(mdates.MonthLocator(interval=2))  # Set the spacing of the x-axis major ticks

plt.gcf().autofmt_xdate()  # Automatically rotate date labels

"""Set the format of the axes"""
# ax1=ax1.twinx()
# ax1=plt.twinx()
ax2 = ax1.twinx()
# Set major ticks, one tick per month
fmt_half_year = mdates.MonthLocator(interval=1)
ax1.xaxis.set_major_locator(fmt_half_year)

# # Set minor ticks, one tick per week
# fmt_week = mdates.WeekdayLocator() # Default is fine
# ax1.xaxis.set_major_locator(fmt_week)

# Set the tick format of the x-axis
ax1.xaxis.set_major_formatter(mdates.DateFormatter("%Y-%m"))
# Parameter pad is used to set the distance between ticks and labels
ax1.tick_params(pad=10,labelsize=20)
ax2.tick_params(pad=10,labelsize=20)

df3 = pd.read_csv('freqvitd.csv')
x = df3['created_at']
x = pd.to_datetime(x)               # Convert to date, otherwise the date settings below won't take effect
y3 = df3['d']
ln3, =ax1.plot(x,y3,color="#8491B4",linewidth=4, ls= '-', marker='v',markersize=8)

df4 = pd.read_csv('freqrem.csv')
x = df4['created_at']
x = pd.to_datetime(x)               # Convert to date, otherwise the date settings below won't take effect
y4 = df4['d']
ln4, =ax1.plot(x,y4,color="#E64B35",linewidth=4,ls='-', marker='s', markersize=8)

df5 = pd.read_csv('freqzinc.csv')
x = df5['created_at']
x = pd.to_datetime(x)               # Convert to date, otherwise the date settings below won't take effect
y5 = df5['d']
ln5, =ax1.plot(x,y5,color="#7E6148",linewidth=4,ls='-', marker= 'x', markersize=8)

df6 = pd.read_csv('case_week.csv')
x2 = df6['Date']
x2 = pd.to_datetime(x2)               # Convert to date, otherwise the date settings below won't take effect
y6 = df6['Cases']
ln6, =ax2.plot(x2,y6,color="#3C5488",linewidth=4,ls='-', marker= 'o', markersize=8)

# ax.fill_between(x, y3,color="#8491B4")
# ax.fill_between(x, y1,color="#91D1C2")
# ax.fill_between(x, y2,y3,color="#F39B7F")
# Display the image
# plt.stackplot(x,y3,y1,y2,labels=['vitd','ivermectin','hcq'],colors=["#91D1C2","#F39B7F","#8491B4"],alpha=0.8)

ln1, =ax1.plot(x,y1,color="#91D1C2",linewidth=4,ls='-', marker='o',markersize=8)
ln2, =ax1.plot(x,y2,color="#F39B7F",linewidth=4,ls='-', marker= '^',markersize=8)
ax1.legend(handles=[ln1, ln2,ln4,ln5,ln3],labels=['Ivermectin','Hydroxychloroquine','Remdesivir','Zinc','Vitamin D'],loc='upper left',prop=font1)
ax2.legend(handles=[ln6],labels=['Case number'],loc='upper right',prop=font1)
ax1.set_xlabel('Time',font1)
ax1.set_ylabel('Occurrence',font1)
ax2.set_ylabel('Case number',font1)

# Add annotations
x0 = pd.to_datetime('2020-04-12')   
y0 = 20000/4
ax1.scatter(x0,y0,s=50,color="#3C5488")
ax1.plot([x0,x0],[y0,0],'k--',lw=1.5)
ax1.annotate(r'Trump promotes HCQ', xy=(x0,y0),xycoords='data',xytext=(-10,+30),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x1 = pd.to_datetime('2020-08-02')   
y1 = 30000
ax1.scatter(x1,y1,s=50,color="#3C5488")
ax1.plot([x1,x1],[y1,0],'k--',lw=1.5)
ax1.annotate(r'Epidemiologist openly recommend HCQ', xy=(x1,y1),xycoords='data',xytext=(-10,+10),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x2 = pd.to_datetime('2020-05-24')   
y2 = 42000/4
ax1.scatter(x2,y2,s=50,color="#3C5488")
ax1.plot([x2,x2],[y2,0],'k--',lw=1.5)
ax1.annotate(r'Trump took HCQ to prevent COVID-19', xy=(x2,y2),xycoords='data',xytext=(0,+30),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x4 = pd.to_datetime('2020-05-03')   
y4 = 13000/4
ax1.scatter(x4,y4,s=50,color="#3C5488")
ax1.plot([x4,x4],[y4,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved an EUA to Remdesivir', xy=(x4,y4),xycoords='data',xytext=(-30,+5),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))
             
x5 = pd.to_datetime('2020-10-25')   
y5 = 12000/4
ax1.scatter(x5,y5,s=50,color="#3C5488")
ax1.plot([x5,x5],[y5,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved Remdesivir', xy=(x5,y5),xycoords='data',xytext=(-30,+30),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x6 = pd.to_datetime('2021-04-25')   
# y6 = 13000
# ax1.scatter(x6,y6,s=50,color="#3C5488")
# ax1.plot([x6,x6],[y6,0],'k--',lw=1.5)
# ax1.annotate(r'FDA clarified that Remdesivir was not approved', xy=(x6,y6),xycoords='data',xytext=(-30,+20),textcoords='offset points',
#              fontsize=20)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x7 = pd.to_datetime('2021-08-29')   
y7 = 20500
ax1.scatter(x7,y7,s=50,color="#3C5488")
ax1.plot([x7,x7],[y7,0],'k--',lw=1.5)
ax1.annotate(r'FDA denounced Ivermectin', xy=(x7,y7),xycoords='data',xytext=(-250,-40),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.21

x8 = pd.to_datetime('2021-09-05')   
y8 = 32000
ax1.scatter(x8,y8,s=50,color="#3C5488")
ax1.plot([x8,x8],[y8,0],'k--',lw=1.5)
ax1.annotate(r'CDC warned ivermectin', xy=(x8,y8),xycoords='data',xytext=(-260,-40),textcoords='offset points',
             fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.26

# Add annotations
x10 = pd.to_datetime('2022-01-15')   
# y10 = 32000
y10 = 16000
# ax1.scatter(x10,y10,s=50,color="#3C5488")
# ax1.plot([x10,x10],[y10,0],'k--',lw=1.5)
ax1.annotate(r'Omicron, South Africa', xy=(x10,y10),xycoords='data',xytext=(+30,+30),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x11 = pd.to_datetime('2021-01-03')   
# y11 = 1818837/20
# ax1.scatter(x11,y11,s=50,color="#3C5488")
# ax1.plot([x11,x11],[y11,0],'k--',lw=1.5)
# ax1.annotate(r'Beta, South Africa ', xy=(x11,y11),xycoords='data',xytext=(-10,0),textcoords='offset points',
#              fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x12 = pd.to_datetime('2020-12-27')   
y12 = 4000
# ax1.scatter(x12,y12,s=50,color="#3C5488")
# ax1.plot([x12,x12],[y12,0],'k--',lw=1.5)
ax1.annotate(r'Delta, India', xy=(x12,y12),xycoords='data',xytext=(+10,+200),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# Display the image and save it
plt.savefig('4.tif', bbox_inches = 'tight',dpi=300)
plt.show()# Label the Omicron event
