==> 7_sentidis.ipynb <==
### Calculate the proportion of different sentiments in weekly tweets about various drugs.
#Calculate the weekly proportion of positive tweets about zinc.
import pandas as pd
drug=pd.read_csv('zinc.csv')
positive=drug[drug['stance']=='positive']
data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
p=data.resample('W', label='right', closed='right').d.sum()
p.to_csv('freqzincp.csv')
#Then divide by the total number of weekly tweets.

#Calculate the weekly proportion of negative tweets about zinc.
import pandas as pd
drug=pd.read_csv('zinc.csv')
negative=drug[drug['stance']=='negative']
data=negative
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
n=data.resample('W', label='right', closed='right').d.sum()
n.to_csv('freqzincn.csv')
#Then divide by the total number of weekly tweets.
#Calculate the weekly proportion of positive tweets about HCQ.
import pandas as pd
drug=pd.read_csv('HCQ.csv')
positive=drug[drug['stance']=='positive']
data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
p=data.resample('W', label='right', closed='right').d.sum()
p.to_csv('freqHCQp.csv')
#Then divide by the total number of weekly tweets.

#Calculate the weekly proportion of negative tweets about HCQ.
import pandas as pd
drug=pd.read_csv('HCQ.csv')
negative=drug[drug['stance']=='negative']
data=negative
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
n=data.resample('W', label='right', closed='right').d.sum()
n.to_csv('freqHCQn.csv')
#Then divide by the total number of weekly tweets.
#### Extract peak tweets to examine discussion hotspots.
import pandas as pd
data=pd.read_csv('/data1/data_8t/ivermectin/sentiment/Remdesivir|velkury.csv')
# positive=drug[drug['stance']=='negative']
# data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
data=data['2021-08-30':'2021-09-05']
data=data.sample(n=200, replace=True, random_state=666)
for i in data['full_text']:
    print(i)
    print('-------------------------')
#Calculate the weekly proportion of positive tweets about rem.
import pandas as pd
drug=pd.read_csv('rem.csv')
positive=drug[drug['stance']=='positive']
data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
p=data.resample('W', label='right', closed='right').d.sum()
p.to_csv('freqremp.csv')
#then divide by the total number of weekly tweets.

#Calculate the weekly proportion of negative tweets about rem.
import pandas as pd
drug=pd.read_csv('rem.csv')
negative=drug[drug['stance']=='negative']
data=negative
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
n=data.resample('W', label='right', closed='right').d.sum()
n.to_csv('freqremn.csv')
#then divide by the total number of weekly tweets.
#Calculate the weekly proportion of positive tweets about vitd.
import pandas as pd
drug=pd.read_csv('vitd.csv')
positive=drug[drug['stance']=='positive']
data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
p=data.resample('W', label='right', closed='right').d.sum()
p.to_csv('freqvitdp.csv')
#then divide by the total number of weekly tweets.

#Calculate the weekly proportion of negative tweets about vitd.
import pandas as pd
drug=pd.read_csv('vitd.csv')
negative=drug[drug['stance']=='negative']
data=negative
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
n=data.resample('W', label='right', closed='right').d.sum()
n.to_csv('freqvitdn.csv')
#then divide by the total number of weekly tweets.
#Calculate the weekly proportion of positive tweets about iver.
import pandas as pd
drug=pd.read_csv('iver.csv')
positive=drug[drug['stance']=='positive']
data=positive
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
p=data.resample('W', label='right', closed='right').d.sum()
p.to_csv('freqiverp.csv')
#then divide by the total number of weekly tweets.

#Calculate the weekly proportion of negative tweets about iver.
import pandas as pd
drug=pd.read_csv('iver.csv')
negative=drug[drug['stance']=='negative']
data=negative
data['created_at'] = pd.to_datetime(data['created_at'])
data.index = data['created_at']
data=data.sort_index()
#del data['created_at']
data.insert(data.shape[1], 'd', 1)
n=data.resample('W', label='right', closed='right').d.sum()
n.to_csv('freqivern.csv')
#then divide by the total number of weekly tweets.
#### Convert the count to a percentage, which means dividing by the total number of tweets.
# Negative Sentiment Proportion
import pandas as pd
drug=['hcq','iver','zinc','vitd','rem']
#drug=['vitd']

for i in drug:
    n=pd.read_csv('freq'+i+"n.csv")
    s=pd.read_csv('freq'+i+"s.csv")
    neg=pd.merge(n, s, on='created_at')
    print(neg.head(1))
    #df_data['average'] =df_data.apply(lambda x: x['total'] / x['num'], axis=1)
    try:
        neg['per'] =neg.apply(lambda x: x['neg_num'] / x['senti_num'], axis=1)
        
    except ZeroDivisionError: # because some users has no text
        neg['per']=0
    neg.to_csv('freq'+i+"np.csv")
# Positive Sentiment Proportion
import pandas as pd
drug=['hcq','iver','zinc','vitd','rem']
#senti=['pos','pos']
for i in drug:
    p=pd.read_csv('freq'+i+"p.csv")
    s=pd.read_csv('freq'+i+"s.csv")
    pos=pd.merge(p, s, on='created_at')
    print(pos.head(1))
    #df_data['average'] =df_data.apply(lambda x: x['total'] / x['num'], axis=1)
    try:
        pos['per'] =pos.apply(lambda x: x['pos_num'] / x['senti_num'], axis=1)   
    except ZeroDivisionError: # because some users has no text
        pos['per']=0
    pos.to_csv('freq'+i+"pp.csv",index=False)
import pandas as pd
drug=['hcq','iver','zinc','vitd','rem']
#senti=['pos','neg']
for j in senti:
    for i in drug:
        p=pd.read_csv('freq'+i+"p.csv")
        s=pd.read_csv('freq'+i+"s.csv")
        j=pd.merge(p, s, on='created_at')
        print(j.head(1))
        #df_data['average'] =df_data.apply(lambda x: x['total'] / x['num'], axis=1)
        try:
            j['per'] =j.apply(lambda x: x['pos_num'] / x['senti_num'], axis=1)
            
        except ZeroDivisionError: # because some users has no text
            j['per']=0
        j.to_csv('freq'+i+j+".csv")
### Plotting
import numpy as np
import pandas as pd
import matplotlib.dates as mdates    #Processing date
import matplotlib.pyplot as plt
#Read date
fig = plt.figure(figsize=(35, 12))
ax1 = fig.add_subplot()
font1 = {'family' : 'DejaVu Sans',
'weight' : 'normal',
'size'   : 20,
}

df1 = pd.read_csv('freqiverpp.csv')
x = df1['created_at']
x = pd.to_datetime(x)               #Change to date, otherwise the date setting below will not take effect
y1 = df1['per']
ln1, =ax1.plot(x,y1,color="#91D1C2",linewidth=4,ls='-', marker='o',markersize=8)


plt.gca().xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))   #Set the main scale display format of the x-axis (date)
plt.gca().xaxis.set_major_locator(mdates.MonthLocator(interval=2))  #Set the interval of the main scale of the x-axis (date)

plt.gcf().autofmt_xdate()  # Rotate the date display angle
 
"""Set the font size of the scale"""
#ax2 = ax1.twinx()
# set main scale, one month one scale
fmt_half_year = mdates.MonthLocator(interval=1)
ax1.xaxis.set_major_locator(fmt_half_year)

# set minor scale, one week one scale
# fmt_week = mdates.WeekdayLocator() # 默认即可
# ax1.xaxis.set_minor_locator(fmt_week)

# set date display format
ax1.xaxis.set_major_formatter(mdates.DateFormatter("%Y-%m"))
#parameter pad is the distance between the axis and the scale
ax1.tick_params(pad=10,labelsize=20)
#ax2.tick_params(pad=10,labelsize=20)

df3 = pd.read_csv('freqvitdpp.csv')
x = df3['created_at']
x = pd.to_datetime(x)               
y3 = df3['per']
ln3, =ax1.plot(x,y3,color="#8491B4",linewidth=4, ls= '-', marker='v',markersize=8)

df2 = pd.read_csv('freqhcqpp.csv')
x = df2['created_at']
x = pd.to_datetime(x)             
y2 = df2['per']
ln2, =ax1.plot(x,y2,color="#F39B7F",linewidth=4,ls='-', marker= '^',markersize=8)

df4 = pd.read_csv('freqrempp.csv')
x = df4['created_at']
x = pd.to_datetime(x)               
y4 = df4['per']
ln4, =ax1.plot(x,y4,color="#E64B35",linewidth=4,ls='-', marker='s',markersize=8 )

df5 = pd.read_csv('freqzincpp.csv')
x = df5['created_at']
x = pd.to_datetime(x)               
y5 = df5['per']
ln5, =ax1.plot(x,y5,color="#7E6148",linewidth=4,ls='-', marker= 'x', markersize=8)



ax1.legend(handles=[ln1, ln2,ln4,ln5,ln3],labels=['Ivermectin','Hydroxychloroquine','Remdesivir','Zinc','Vitamin D'],loc='upper right',prop=font1)
#ax1.legend(handles=[ ln2,ln4,ln5],labels=['HCQ','Remdesivir','Zinc'],loc='upper left',prop=font1)
#ax2.legend(handles=[ln6],labels=['Case number'],loc='upper right',prop=font1)
ax1.set_xlabel('Time',font1)
ax1.set_ylabel('Positive sentiment ratio',font1)
#ax2.set_ylabel('Case number',font1)

x0 = pd.to_datetime('2020-04-12')   
y0 = 0.28391046489382055
ax1.scatter(x0,y0,s=50,color="#3C5488")
ax1.plot([x0,x0],[y0,0],'k--',lw=1.5)
ax1.annotate(r'Trump continued to emphasize the effectiveness of HCQ', xy=(x0,y0),xycoords='data',xytext=(-100,+300),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x1 = pd.to_datetime('2020-08-02')   
# y1 = 0.1
# ax1.scatter(x1,y1,s=50,color="#3C5488")
# ax1.plot([x1,x1],[y1,0],'k--',lw=1.5)
# ax1.annotate(r'Yale epidemiologist says HCQ is the key to defeating COVID-19', xy=(x1,y1),xycoords='data',xytext=(-10,+10),textcoords='offset points',
#              fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x2 = pd.to_datetime('2020-05-24')   
y2 = 0.25826475101129864
ax1.scatter(x2,y2,s=50,color="#3C5488")
ax1.plot([x2,x2],[y2,0],'k--',lw=1.5)
ax1.annotate(r'Trump said he has been taking HCQ to prevent COVID-19', xy=(x2,y2),xycoords='data',xytext=(-80,+180),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x4 = pd.to_datetime('2020-05-03')   
y4 = 0.18105065666041276
ax1.scatter(x4,y4,s=50,color="#3C5488")
ax1.plot([x4,x4],[y4,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved an EUA to use Remdesivir', xy=(x4,y4),xycoords='data',xytext=(-30,-50),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))
             
x5 = pd.to_datetime('2020-10-25')   
y5 =0.17586795674445077
ax1.scatter(x5,y5,s=50,color="#3C5488")
ax1.plot([x5,x5],[y5,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved Remdesivir', xy=(x5,y5),xycoords='data',xytext=(-30,+30),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x6 = pd.to_datetime('2021-04-25')   
# y6 = 13000
# ax1.scatter(x6,y6,s=50,color="#3C5488")
# ax1.plot([x6,x6],[y6,0],'k--',lw=1.5)
# ax1.annotate(r'FDA clarified that Remdesivir was not approved', xy=(x6,y6),xycoords='data',xytext=(-30,+20),textcoords='offset points',
#              fontsize=20)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x7 = pd.to_datetime('2021-08-29')   
y7 = 0.271370999
ax1.scatter(x7,y7,s=50,color="#3C5488")
ax1.plot([x7,x7],[y7,0],'k--',lw=1.5)
ax1.annotate(r'FDA denounced Ivermectin ', xy=(x7,y7),xycoords='data',xytext=(-300,+200),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.21

x7 = pd.to_datetime('2020-12-08')   
y7 = 0.551136364
ax1.scatter(x7,y7,s=50,color="#3C5488")
ax1.plot([x7,x7],[y7,0],'k--',lw=1.5)
ax1.annotate(r'U.S. Sen. Ron Johnson promoted ivermectin on a Senate hearing.', xy=(x7,y7),xycoords='data',xytext=(-200,0),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.21


x8 = pd.to_datetime('2021-09-05')   
y8 = 0.275290916
ax1.scatter(x8,y8,s=50,color="#3C5488")
ax1.plot([x8,x8],[y8,0],'k--',lw=1.5)
ax1.annotate(r'CDC issued a warning about ivermectin', xy=(x8,y8),xycoords='data',xytext=(-310,+120),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))
#Show the plot and save
plt.savefig('pp.tif', bbox_inches = 'tight',transparent=True,dpi=300)
plt.show()
import numpy as np
import pandas as pd
import matplotlib.dates as mdates
import matplotlib.pyplot as plt
fig = plt.figure(figsize=(35, 12))
ax1 = fig.add_subplot()
font1 = {'family' : 'DejaVu Sans',
'weight' : 'normal',
'size'   : 20,
}

df1 = pd.read_csv('freqivernp.csv')
x = df1['created_at']
x = pd.to_datetime(x)              
y1 = df1['per']
ln1, =ax1.plot(x,y1,color="#91D1C2",linewidth=4,ls='-', marker='o',markersize=8)



plt.gca().xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))  
plt.gca().xaxis.set_major_locator(mdates.MonthLocator(interval=2))  

plt.gcf().autofmt_xdate()  
 
#ax2 = ax1.twinx()
fmt_half_year = mdates.MonthLocator(interval=1)
ax1.xaxis.set_major_locator(fmt_half_year)

# fmt_week = mdates.WeekdayLocator()
# ax1.xaxis.set_minor_locator(fmt_week)

ax1.xaxis.set_major_formatter(mdates.DateFormatter("%Y-%m"))
ax1.tick_params(pad=10,labelsize=20)
#ax2.tick_params(pad=10,labelsize=20)

df3 = pd.read_csv('freqvitdnp.csv')
x = df3['created_at']
x = pd.to_datetime(x)             
y3 = df3['per']
ln3, =ax1.plot(x,y3,color="#8491B4",linewidth=4, ls= '-', marker='v',markersize=8)

df2 = pd.read_csv('freqhcqnp.csv')
x = df2['created_at']
x = pd.to_datetime(x)               
y2 = df2['per']
ln2, =ax1.plot(x,y2,color="#F39B7F",linewidth=4,ls='-', marker= '^',markersize=8)

df4 = pd.read_csv('freqremnp.csv')
x = df4['created_at']
x = pd.to_datetime(x)               
y4 = df4['per']
ln4, =ax1.plot(x,y4,color="#E64B35",linewidth=4,ls='-', marker='s',markersize=8 )

df5 = pd.read_csv('freqzincnp.csv')
x = df5['created_at']
x = pd.to_datetime(x)               
y5 = df5['per']
ln5, =ax1.plot(x,y5,color="#7E6148",linewidth=4,ls='-', marker= 'x', markersize=8)



ax1.legend(handles=[ln1, ln2,ln4,ln5,ln3],labels=['Ivermectin','Hydroxychloroquine','Remdesivir','Zinc','Vitamin D'],loc='upper right',prop=font1)
#ax1.legend(handles=[ ln2,ln4,ln5],labels=['HCQ','Remdesivir','Zinc'],loc='upper left',prop=font1)
#ax2.legend(handles=[ln6],labels=['Case number'],loc='upper right',prop=font1)
ax1.set_xlabel('Time',font1)
ax1.set_ylabel('Negative sentiment ratio',font1)
#ax2.set_ylabel('Case number',font1)

x0 = pd.to_datetime('2020-04-12')   
y0 = 0.12110197053759326
ax1.scatter(x0,y0,s=50,color="#3C5488")
ax1.plot([x0,x0],[y0,0],'k--',lw=1.5)
ax1.annotate(r'Trump continued to emphasize the effectiveness of HCQ', xy=(x0,y0),xycoords='data',xytext=(-200,+300),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x1 = pd.to_datetime('2020-08-02')   
# y1 = 0.1
# ax1.scatter(x1,y1,s=50,color="#3C5488")
# ax1.plot([x1,x1],[y1,0],'k--',lw=1.5)
# ax1.annotate(r'Yale epidemiologist says HCQ is the key to defeating COVID-19', xy=(x1,y1),xycoords='data',xytext=(-10,+10),textcoords='offset points',
#              fontsize=25)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x2 = pd.to_datetime('2020-05-24')   
y2 = 0.23929418328916166
ax1.scatter(x2,y2,s=50,color="#3C5488")
ax1.plot([x2,x2],[y2,0],'k--',lw=1.5)
ax1.annotate(r'Trump said he has been taking HCQ to prevent COVID-19', xy=(x2,y2),xycoords='data',xytext=(-50,+100),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x4 = pd.to_datetime('2020-05-03')   
y4 = 0.11163227016885553
ax1.scatter(x4,y4,s=50,color="#3C5488")
ax1.plot([x4,x4],[y4,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved an EUA to use Remdesivir', xy=(x4,y4),xycoords='data',xytext=(-30,100),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))
             
x5 = pd.to_datetime('2020-10-25')   
y5 =0.1462720546385885
ax1.scatter(x5,y5,s=50,color="#3C5488")
ax1.plot([x5,x5],[y5,0],'k--',lw=1.5)
ax1.annotate(r'FDA approved Remdesivir', xy=(x5,y5),xycoords='data',xytext=(+100,+100),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

# x6 = pd.to_datetime('2021-04-25')   
# y6 = 13000
# ax1.scatter(x6,y6,s=50,color="#3C5488")
# ax1.plot([x6,x6],[y6,0],'k--',lw=1.5)
# ax1.annotate(r'FDA clarified that Remdesivir was not approved', xy=(x6,y6),xycoords='data',xytext=(-30,+20),textcoords='offset points',
#              fontsize=20)#,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x7 = pd.to_datetime('2021-08-29')   
y7 = 0.28629001
ax1.scatter(x7,y7,s=50,color="#3C5488")
ax1.plot([x7,x7],[y7,0],'k--',lw=1.5)
ax1.annotate(r'FDA denounced Ivermectin ', xy=(x7,y7),xycoords='data',xytext=(-100,+100),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.21

x7 = pd.to_datetime('2020-12-08')   
y7 = 0.034090909
ax1.scatter(x7,y7,s=50,color="#3C5488")
ax1.plot([x7,x7],[y7,0],'k--',lw=1.5)
ax1.annotate(r'U.S. Sen. Ron Johnson promoted ivermectin on a Senate hearing.', xy=(x7,y7),xycoords='data',xytext=(-200,+10),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))#2021.8.21


x8 = pd.to_datetime('2021-09-05')   
y8 = 0.265812688
ax1.scatter(x8,y8,s=50,color="#3C5488")
ax1.plot([x8,x8],[y8,0],'k--',lw=1.5)
ax1.annotate(r'CDC issued a warning about ivermectin', xy=(x8,y8),xycoords='data',xytext=(-90,+50),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

x8 = pd.to_datetime('2021-09-05')   
y8 = 0.265812688
ax1.scatter(x8,y8,s=50,color="#3C5488")
ax1.plot([x8,x8],[y8,0],'k--',lw=1.5)
ax1.annotate(r'CDC issued a warning about ivermectin', xy=(x8,y8),xycoords='data',xytext=(-90,+50),textcoords='offset points',
             fontsize=25,arrowprops=dict(arrowstyle='->',connectionstyle='arc3,rad=.2'))

plt.savefig('nn.tif', bbox_inches = 'tight',transparent=True,dpi=300)
plt.show()
### Evaluate the model
#Randomly select 50 and evaluate
import pandas as pd
df = pd.read_csv("zinc.csv",lineterminator='\n',usecols=['id_str','created_at','user','full_text'],low_memory=False, keep_default_na=False)
import random
alist = random.sample(range(1,len(df['full_text'])),10)
print(df.reset_index(drop=True))
for i in alist:
    print(df['full_text','stance'][i])
    print('-------------------------------------------')
#Randomly select 50 and evaluate
import pandas as pd
iver = pd.read_csv("iver.csv",lineterminator='\n',low_memory=False, keep_default_na=False)
iver20=iver.sample(n=20,random_state=1)
iver20.to_csv('iver20.csv')

rem = pd.read_csv("rem.csv",lineterminator='\n',low_memory=False, keep_default_na=False)
rem20=rem.sample(n=20,random_state=1)
rem20.to_csv('rem20.csv')

HCQ = pd.read_csv("HCQ.csv",lineterminator='\n',low_memory=False, keep_default_na=False)
HCQ20=HCQ.sample(n=20,random_state=1)
HCQ20.to_csv('HCQ20.csv')

vitd = pd.read_csv("vitd.csv",lineterminator='\n',low_memory=False, keep_default_na=False)
vitd20=vitd.sample(n=20,random_state=1)
vitd20.to_csv('vitd20.csv')

zinc = pd.read_csv("zinc.csv",lineterminator='\n',low_memory=False, keep_default_na=False)
zinc20=zinc.sample(n=20,random_state=1)
zinc20.to_csv('zinc20.csv')

df=pd.concat([iver20,HCQ20,vitd20,rem20,zinc20])
df.to_csv('random100.csv')
#### Overall Sentiment Distribution Figure
import numpy as np
import matplotlib.pyplot as plt
plt.figure(figsize=(11,5))
color = ["#E64B35","#4DBBD5","#00A087","#3C5488","#F39B7F","#8491B4","#91D1C2","#DC0000","#7E6148"]
labels = ['Ivermectin','Hydroxychloroquine','Remdesivir','Zinc','Vitamin D']
first = [0.361036609,0.382484138,0.12766002,0.452212509,0.45034322] #统计出来的数字
second = [0.451497828,0.48150596,0.677235151,0.535742938,0.530308739]
third = [0.187465563,0.136009902,0.195104829,0.012044552,0.019348041]
data = [first, second,third ]

x = range(len(labels))
width = 0.35

bottom_y = np.zeros(len(labels))
data = np.array(data)
sums = np.sum(data, axis=0)
j=0
bar1=plt.bar(x, first, width, bottom=bottom_y,color=color[0])
bar2=plt.bar(x, second, width, bottom=bottom_y+first,color=color[1])
bar3=plt.bar(x, third, width, bottom=bottom_y+first+second,color=color[2])

plt.xticks(x, labels,fontsize=15)
plt.yticks(fontsize=15)
#plt.title('Drug sentiment analysis',fontsize=15)
#plt.xlabel('Drug entity',fontsize=15)
plt.ylabel('Sentiment percentage',fontsize=15)

plt.legend(handles=[bar3,bar2,bar1],labels=["Negative","Neutral","Positive"],loc=2, bbox_to_anchor=(1.05,1.0),fontsize=15,borderaxespad = 0.)
#plt.gcf().subplots_adjust(right=0.5)
plt.savefig('total_senti.tif',bbox_inches = 'tight',transparent=True,dpi=300)

plt.show()
